Fix benchmarks, update TestClient to support server ssl, move common parts to pytest_plugin.py
